consolidate code, process announcements in command response as well
