+++ model/command_response.go
@@ -5,7 +5,6 @@ package model
 
 import (
 	"encoding/json"
-	"fmt"
 	"io"
 )
 
@@ -40,26 +39,7 @@ func CommandResponseFromJson(data io.Reader) *CommandResponse {
 		return nil
 	}
 
-	// Ensure attachment fields are stored as strings
-	var nonNilAttachments []*SlackAttachment
-	for _, attachment := range o.Attachments {
-		if attachment == nil {
-			continue
-		}
-		nonNilAttachments = append(nonNilAttachments, attachment)
-		var nonNilFields []*SlackAttachmentField
-		for _, field := range attachment.Fields {
-			if field == nil {
-				continue
-			}
-			nonNilFields = append(nonNilFields, field)
-			if field.Value != nil {
-				field.Value = fmt.Sprintf("%v", field.Value)
-			}
-		}
-		attachment.Fields = nonNilFields
-	}
-	o.Attachments = nonNilAttachments
+	processAttachments(&o.Attachments)
 
 	return &o
 }
