+++ model/slack_attachment.go
@@ -3,6 +3,11 @@
 
 package model
 
+import (
+	"fmt"
+	"strings"
+)
+
 type SlackAttachment struct {
 	Id         int64                   `json:"id"`
 	Fallback   string                  `json:"fallback"`
@@ -26,4 +31,49 @@ type SlackAttachmentField struct {
 	Title string      `json:"title"`
 	Value interface{} `json:"value"`
 	Short bool        `json:"short"`
+}
+
+// To mention @channel via a webhook in Slack, the message should contain
+// <!channel>, as explained at the bottom of this article:
+// https://get.slack.help/hc/en-us/articles/202009646-Making-announcements
+func expandAnnouncement(text string) string {
+	c1 := "<!channel>"
+	c2 := "@channel"
+	if strings.Contains(text, c1) {
+		return strings.Replace(text, c1, c2, -1)
+	}
+	return text
+}
+
+// Expand announcements in incoming webhooks from Slack. Those announcements
+// can be found in the text attribute, or in the pretext, text, title and value
+// attributes of the attachment structure. The Slack attachment structure is
+// documented here: https://api.slack.com/docs/attachments
+func processAttachments(attachments *[]*SlackAttachment) {
+	var nonNilAttachments []*SlackAttachment
+	for _, attachment := range *attachments {
+		if attachment == nil {
+			continue
+		}
+		nonNilAttachments = append(nonNilAttachments, attachment)
+
+		attachment.Pretext = expandAnnouncement(attachment.Pretext)
+		attachment.Text = expandAnnouncement(attachment.Text)
+		attachment.Title = expandAnnouncement(attachment.Title)
+
+		var nonNilFields []*SlackAttachmentField
+		for _, field := range attachment.Fields {
+			if field == nil {
+				continue
+			}
+			nonNilFields = append(nonNilFields, field)
+
+			if field.Value != nil {
+				// Ensure the value is set to a string if it is set
+				field.Value = expandAnnouncement(fmt.Sprintf("%v", field.Value))
+			}
+		}
+		attachment.Fields = nonNilFields
+	}
+	*attachments = nonNilAttachments
 }
